+++ tests/test_gsheet_base.py
@@ -1,4 +1,4 @@
-from pyfreeleh.google.sheet.base import A1Range, CellSelector
+from pyfreeleh.providers.google.sheet.base import A1Range, CellSelector
 
 
 def test_cell_selector():
